+++ src/main/java/com/acm/app/user/dao/UserDAO.java
@@ -1,22 +1,22 @@
 package com.acm.app.user.dao;
 
-import static com.acm.app.user.mapper.UserMapper.USER_MAPPER;
-import static com.acm.app.user.mapper.UserCredentialMapper.USER_CREDENTIAL_MAPPER;
-import static com.acm.service.sql.SqlClient.getPage;
-import static com.acm.service.sql.SqlClient.getTemplate;
+import com.acm.app.user.client.domain.User;
+import com.acm.app.user.client.domain.request.UserGetRequest;
+import com.acm.service.exceptions.UserNotFoundException;
+import com.acm.service.sql.SQLBuilder;
+import com.google.common.collect.Sets;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
-
-import com.acm.app.user.client.domain.User;
-import com.acm.app.user.client.domain.request.UserGetRequest;
-import com.acm.service.sql.SQLBuilder;
-import com.google.common.collect.Sets;
+import static com.acm.app.user.mapper.UserCredentialMapper.USER_CREDENTIAL_MAPPER;
+import static com.acm.app.user.mapper.UserMapper.USER_MAPPER;
+import static com.acm.service.sql.SqlClient.getPage;
+import static com.acm.service.sql.SqlClient.getTemplate;
 
 /**
  * Used to manage queries about users from the database
@@ -36,7 +36,8 @@ public class UserDAO {
 	 * @param request {@link UserGetRequest} object to search upon
 	 * @return List of {@link User}
 	 */
-	public List<User> getUser(UserGetRequest request) {
+	public List<User> getUser(UserGetRequest request)
+	{
 		Map<String, Set<?>> params = new HashMap<>();
 
 		sqlBuilder.setQueryFile("userDAO");
@@ -45,7 +46,7 @@ public class UserDAO {
 		return getPage(sqlBuilder.getSql("getUsers"), USER_MAPPER);
 	}
 	
-	public User getUserCredentials(UserGetRequest request)
+	public User getUserCredentials(UserGetRequest request) throws UserNotFoundException
 	{
 		Map<String, Set<?>> params = new HashMap<>();
 		
@@ -53,6 +54,10 @@ public class UserDAO {
 			params.put("userId", Sets.newHashSet(request.getId()));
 		} else {
 			params.put("username", Sets.newHashSet(request.getUsername()));
+			if(request.getUsername() == null)
+			{
+				throw new UserNotFoundException("User not found at:", request.getId());
+			}
 		}
 
 		sqlBuilder.setQueryFile("userDAO");