+++ src/main/java/com/acm/app/user/service/UserService.java
@@ -1,13 +1,13 @@
 package com.acm.app.user.service;
 
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
-
 import com.acm.app.user.client.domain.User;
 import com.acm.app.user.client.domain.request.UserGetRequest;
 import com.acm.app.user.dao.UserDAO;
+import com.acm.service.exceptions.UserNotFoundException;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+import java.util.List;
 
 /**
  * UserService sends request to userDAO
@@ -31,7 +31,7 @@ public class UserService {
 		return userDAO.getUser(request);
 	}
 	
-	public User getUserCredentials(UserGetRequest request)
+	public User getUserCredentials(UserGetRequest request) throws UserNotFoundException
 	{
 		return userDAO.getUserCredentials(request);
 	}