+++ src/main/java/com/acm/service/exceptions/InvalidPasswordException.java
@@ -1,0 +1,10 @@
+package com.acm.service.exceptions;
+
+public class InvalidPasswordException extends Exception
+{
+    private static final long serialVersionUID = -1L;
+    public InvalidPasswordException()
+    {
+        super("Entered password is incorrect. Please try again.");
+    }
+}