+++ src/main/java/com/acm/service/exceptions/UserNotFoundException.java
@@ -1,0 +1,9 @@
+package com.acm.service.exceptions;
+
+public class UserNotFoundException extends Exception
+{
+    private static final long serialVersionUID = -1L;
+    public UserNotFoundException(String message, int id){
+        super("User not found at:"+ id);
+    }
+}